extract util-function for assets
