+++ src/modules/olMap/formats/kml.js
@@ -4,7 +4,8 @@ import { Icon, Style } from 'ol/style';
 import CircleStyle from 'ol/style/Circle';
 import { KML } from 'ol/format';
 import { $injector } from '../../../injection';
-import { AssetSourceType, getAssetSource } from '../utils/olStyleUtils';
+import { AssetSourceType, getAssetSource } from '../../../utils/assets';
+
 
 export const KML_PROJECTION_LIKE = 'EPSG:4326';
