+++ src/modules/toolbox/components/drawToolContent/DrawToolContent.js
@@ -7,8 +7,8 @@ import css from './drawToolContent.css';
 import { StyleSizeTypes } from '../../../../services/domain/styles';
 import { clearDescription, clearText, finish, remove, reset, setDescription, setStyle, setType } from '../../../../store/draw/draw.action';
 import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
-import { AssetSourceType, getAssetSource } from '../../../olMap/utils/olStyleUtils';
 import { hexToRgb } from '../../../../utils/colors';
+import { AssetSourceType, getAssetSource } from '../../../../utils/assets';
 
 const Update = 'update';
 const Update_Tools = 'update_tools';
